% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.tps_travail.R
\name{app.tps_travail}
\alias{app.tps_travail}
\title{Application suivi temps de travail}
\usage{
app.tps_travail()
}
\description{
Cette fonction permet de lancer l'application shiny de suivi du temps de travail
}
\examples{
app.tps_travail()
}
