% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associatif.mandats.R
\name{associatif.mandats}
\alias{associatif.mandats}
\title{Export des mandats}
\usage{
associatif.mandats(
  mandat = c("Tous", "Président", "Trésorier", "Garde particulier", "Secrétaire",
    "Vice-président", "Administrateur", "Délégué"),
  sapl = "Toutes",
  formatage = c("Complet", "Propre"),
  export = T
)
}
\arguments{
\item{mandat}{Mandat recherché : Tous (par défault), Président, Trésorier, Garde particulier, Délégué, Secrétaire, Vice-président, Administrateur}

\item{sapl}{SAPL recherchée : Toutes (par défault), ou nom officiel de la SAPL}

\item{formatage}{Type de formatage de la sortie : \code{Complet} (par défault), \code{Propre}}

\item{export}{Si \code{TRUE} (par défault), exporte les mandats au format excel. Si \code{FALSE}, ne l'exporte pas.}
}
\description{
Cette fonction permet d'exporter les mandats des SAPL
}
\examples{
associatif.mandats()
associatif.mandats("Président", "Toutes")
associatif.mandats("Président", "La Truite de l'Orbe")
}
\keyword{formatage}
