% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte.departement.R
\name{carte.departement}
\alias{carte.departement}
\title{Création d'une carte départementale}
\usage{
carte.departement(
  occupation_sols = T,
  routes_principales = T,
  voies_ferrees = T,
  communes_principales = F,
  zoom = 8
)
}
\arguments{
\item{occupation_sols}{Affichage de l'occupation des sols en couleur (\code{TRUE} par défaut)}

\item{routes_principales}{Affichage des routes principales (\code{TRUE} par défaut)}

\item{voies_ferrees}{Affichage des voies ferrées (\code{TRUE} par défaut)}

\item{communes_principales}{Affichage des communes principales (\code{FALSE} par défaut)}

\item{zoom}{Niveau de zoom du MNT (\code{8} par défaut)}
}
\description{
Cette fonction permet de créer un fond de carte départementale au format ggplot2
}
\examples{
carte.departement()
carte.departement(routes_principales = F, voies_ferrees = F)
}
\keyword{cartographie}
