% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte.territoire.R
\name{carte.territoire}
\alias{carte.territoire}
\title{Création d'un fond de carte}
\usage{
carte.territoire(
  emprise = NA,
  occupation_sols = T,
  routes_principales = T,
  voies_ferrees = T,
  communes_principales = F,
  altitude = T,
  zoom = 8,
  position_nord = "tl",
  position_echelle = "tr"
)
}
\arguments{
\item{emprise}{Objet sf (polygone) d'emprise de la carte à représenter}

\item{occupation_sols}{Affichage de l'occupation des sols en couleur (\code{TRUE} par défaut)}

\item{routes_principales}{Affichage des routes principales (\code{TRUE} par défaut)}

\item{voies_ferrees}{Affichage des voies ferrées (\code{TRUE} par défaut)}

\item{communes_principales}{Affichage des communes principales (\code{FALSE} par défaut)}

\item{altitude}{Affichage d'un fond altitudinal (\code{TRUE} par défaut)}

\item{zoom}{Niveau de zoom du MNT (\code{8} par défaut)}

\item{position_nord}{Position de la flèche du nord (\code{tl} par défaut)}

\item{position_echelle}{Position de l'échelle (\code{tr} par défaut)}
}
\description{
Cette fonction permet de créer un fond de carte départemental ou local au format ggplot2
}
\examples{
carte.territoire()
carte.territoire(emprise, altitude = F)
carte.territoire(routes_principales = F, voies_ferrees = F)
}
\keyword{cartographie}
