% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.commentaires.R
\name{chronique.commentaires}
\alias{chronique.commentaires}
\title{Listage des commentaires de chroniques}
\usage{
chronique.commentaires(
  x = NA_character_,
  Recherche = c("Station", "Année biologique")
)
}
\arguments{
\item{Recherche}{Type de données recherchées}

\item{Nom}{Nom recherché}
}
\description{
Cette fonction permet de lister les commentaires de la BDD Chroniques
}
\examples{
chronique.commentaires("HER0-6","Station")
chronique.commentaires("2017","Année biologique")
Commentaires <- Stations$chsta_coderhj \%>\% map_dfr(~ chronique.commentaires(., Recherche = "Station"))
}
\keyword{chronique}
