% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.DCE.R
\name{chronique.DCE}
\alias{chronique.DCE}
\title{Réaliser un export au format DCE AE-OFB de données thermiques}
\usage{
chronique.DCE(
  data = data,
  projet = NA_character_,
  export = T,
  dep39 = F,
  fichierStations = NA_character_,
  fichierSuivis = NA_character_,
  fichierCapteurs = NA_character_
)
}
\arguments{
\item{data}{Data.frame issu de chronique.mesures, ne pouvant pas contenir différentes stations}

\item{projet}{Nom du projet}

\item{export}{Si \code{TRUE}, exporte les résultats (\code{FALSE} par défaut)}

\item{dep39}{Si \code{FALSE} (par défault), ne va pas rechercher les données de stations dans la base locale et donc export simplifié. Si \code{TRUE}, fait la jointure SIG. Possibilité d'utiliser \code{autre} afin de sélectionner un fichier source de stations}

\item{fichierStations}{Permet de passer automatiquement la localisation du fichier de stations si on le souhaite, sinon le demandera}

\item{fichierSuivis}{Permet de passer automatiquement la localisation du fichier de suivis si on le souhaite, sinon le demandera}

\item{fichierCapteurs}{Permet de passer automatiquement la localisation du fichier des capteurs si on le souhaite, sinon le demandera}
}
\description{
Permet de réaliser un export de données thermiques au format DCE AE-OFB
}
\examples{
chronique.DCE(data)
chronique.mesures("SUR0-9", "Thermie") \%>\% chronique.DCE()
mesures_annuelles \%>\% filter(chmes_validation == "Validé") \%>\% group_split(chmes_coderhj) \%>\% map(~ chronique.DCE(data =., projet = glue("{today()}_{projet}"), dep39 = T))
}
\keyword{chronique}
