% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.mesures.R
\name{chronique.mesures}
\alias{chronique.mesures}
\title{Extraction de données brutes de chronique}
\usage{
chronique.mesures(
  CodeStation = character(0),
  Type = c("Thermie", "Piézométrie", "Tout", "Hydrologie", "Oxygénation",
    "Pluviométrie"),
  annee = numeric(0),
  Valide = T
)
}
\arguments{
\item{CodeStation}{Code RHJ de la station}

\item{Type}{Type de chronique recherchée : Thermie (par défaut), piézométrie, Hydrologie, Oxygénation, pluviométrie ou toutes les données}

\item{annee}{Année biologique recherchée si spécifiée, sinon ensemble de la chronique (par défaut)}

\item{Valide}{Validité des données (\code{TRUE} par défaut)}
}
\description{
Permet d'extraire les données de chronique d'une station
}
\examples{
chronique.mesures("HER0-6", "Thermie") 
chronique.mesures("DRO6-8", "Thermie", "2013")
chronique.mesures("GCLzhaval", "Piézométrie", "2015", Valide = F)
chronique.mesures("GCLzhaval", "Tout", "2015", Valide = F)
Mesures <- Stations \%>\% group_split(chsta_coderhj) \%>\% purrr::map_dfr(~ chronique.mesures(.$chsta_coderhj, "Thermie"))
c("GLA4-4", "GLA6-9") \%>\% purrr::map_dfr(~ chronique.mesures(., "Tout", Valide = F))
}
\keyword{chronique}
