% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.suivi.R
\name{chronique.suivi}
\alias{chronique.suivi}
\title{Extraction suivi terrain chroniques}
\usage{
chronique.suivi(
  x = NA_character_,
  Recherche = c("MO", "Opérateur", "Station", "Date", "Capteur"),
  Sortie = c("Complet", "Propre", "Simple")
)
}
\arguments{
\item{x}{Variable dont on cherche le suivi (MO, opérateur, Station, Date, Capteur)}

\item{Recherche}{Type de donnée de suivi}

\item{Sortie}{Forme du dataframe de sortie - \code{Complet} (par défault - tous les champs), \code{Propre} ou \code{Simple}}
}
\description{
Cette fonction permet d'extraire des données de suivi de terrain
}
\examples{
chronique.suivi("CG39", Recherche = "MO")
chronique.suivi("JB-Stéphane", Recherche = "Opérateur")
chronique.suivi("DRO14-2", Recherche = "Station")
chronique.suivi("2015-10-23", Recherche = "Date")
chronique.suivi("9759803", Recherche = "Capteur", Sortie = "Propre")
SuiviTerrain <- Stations$chsta_coderhj \%>\% map_dfr(~ chronique.suivi(., Recherche = "Station"))
}
