% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.traitement.R
\name{chronique.traitement}
\alias{chronique.traitement}
\title{Exécuter le traitement de données de chroniques}
\usage{
chronique.traitement(
  data = data,
  projet = NA_character_,
  export = T,
  exportfigures = T,
  exportDCE = F,
  filtrage = F,
  filtrageanneevmm = TRUE,
  filtragedatefperiode = "07-15",
  filtragenbj = 75,
  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique",
    "Barométrie", "Piézométrie", "Piézométrie brute", "Piézométrie compensée",
    "Piézométrie calée", "Piézométrie NGF", "Oxygénation", "Hydrologie",
    "Pluviométrie"),
  dep39 = c(FALSE, TRUE, "autre"),
  archivage = c("Aucun", "Partiel", "Complet"),
  style = c("boxplot", "violon"),
  log = c("Simple", "Verbeux")
)
}
\arguments{
\item{data}{Data.frame issu de chronique.mesures, pouvant contenir différentes stations}

\item{projet}{Nom du projet}

\item{export}{Si \code{TRUE} (par défault), exporte les résultats/figures. Si \code{FALSE}, ne les exporte pas.}

\item{exportfigures}{Si \code{TRUE} (par défaut), exporte les figures. Si \code{FALSE}, ne les exporte pas.}

\item{exportDCE}{Si \code{TRUE}, exporte les mesures au format DCE. Si \code{FALSE} (par défault), ne les exporte pas.}

\item{filtrage}{Si \code{TRUE}, filtre les résultats avec chronique.resultats.filtrage pour ne conserver que les pertinents pour les représentations graphiques inter-annuelles/inter-sites. \code{FALSE} (par défault)}

\item{filtrageanneevmm}{Si \code{TRUE} (par défault), ne conserve que les résultats dont l'année de fin de vmm30j est égale à l'année biologique du résultat}

\item{filtragedatefperiode}{Ne conserver que les résultats dont la date de fin de période est postérieure à la date choisie ("07-15", soit le 15 juillet par défault)}

\item{filtragenbj}{Nombre de journées minimales pour chaque résultat à conserver (75 par défault).}

\item{typemesure}{Défini le type de données et modifie le traitement en fonction}

\item{dep39}{Si \code{FALSE} (par défault), ne va pas rechercher les données de stations dans la base locale et donc export simplifié. Si \code{TRUE}, fait la jointure SIG. Possibilité d'utiliser \code{autre} afin de sélectionner un fichier source de stations}

\item{archivage}{Si \code{Aucun} (par défault), ne va pas créer une archives .zip du répertoire de sortie. Si \code{Partiel}, créé une archive et conserve le répertoire. Si \code{Complet}, créé une archive et supprimer le répertoire.}

\item{log}{Si \code{Simple} (par défault), le log ne sera pas très bavard. Si \code{Verbeux}, le log incluera tous les retours du tidyverse.}
}
\description{
Permet d'exécuter plusieurs opérations de traitement des données de chronique d'une station
}
\examples{
chronique.traitement(data)
chronique.traitement(data, export = F)
DataTravail <- chronique.mesures("SUR0-9", "Thermie") \%>\% chronique.traitement()
DataTravail <- chronique.mesures("SUR0-9", "Thermie") \%>\% chronique.traitement(., export = F, filtrage = T)
DataTravail <- tbl(db,"Mesures") \%>\% filter(CodeRDT == "DOU393-2" | CodeRDT == "DOU394-5") \%>\% collect() \%>\% chronique.traitement()
}
\keyword{chronique}
