% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.abreviation.R
\name{formatage.abreviation}
\alias{formatage.abreviation}
\title{Exportation d'abréviations}
\usage{
formatage.abreviation(
  thematique = c("Complet", "Écosystème", "Espèce", "Chronique", "Biologie",
    "Physico-chimie", "Morphologie", "Document", "Institution", "Usage"),
  formatage = c("Aucun", "Propre", "Latex"),
  export = F,
  exportformat = c(".xlsx", ".pdf", ".tex")
)
}
\arguments{
\item{thematique}{Thématique recherchée : Complet (par défault), Écosystème, Espèce, Biologie, Chronique, Physico-chimie, Morphologie, Document, Institution, Usage}

\item{formatage}{Type de formatage de la sortie : \code{Aucun} (par défault), \code{Propre} ou \code{Latex}}

\item{export}{Si \code{TRUE}, exporte le glossaire. Si \code{FALSE} (par défault), ne l'exporte pas.}

\item{exportformat}{Format d'exportation des acronymes : \code{.xlsx} (par défault), \code{.pdf} ou \code{.tex}.}
}
\description{
Cette fonction permet d'exporter la définition des acronymes
}
\examples{
formatage.abreviation()
formatage.abreviation(formatage = "Latex", export = T)
formatage.abreviation(thematique = "Physico-chimie", formatage = "Propre", export = T)

NomsColonnes <- colnames(Operations)
acronymes <- formatage.abreviation()
Operations <-
  Operations \%>\%
  left_join(acronymes, by = c(Station = "Acronyme")) \%>\%
  select(-Station) \%>\%
  rename(Station = Definition)
Operations <-
  Operations \%>\%
  select(match(NomsColonnes,names(.)))
}
\keyword{glossaire}
