% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.mo.libelle.R
\name{formatage.mo.libelle}
\alias{formatage.mo.libelle}
\title{Reformatage des id en libellés}
\usage{
formatage.mo.libelle(data, colonne_entree = NA_character_, conservation = F)
}
\arguments{
\item{data}{Dataframe contenant les données à transformer}

\item{colonne_entree}{Champ contenant la donnée à ré-écrire}

\item{conservation}{Conservation du champs colonne_entree dans la sortie (avec suffixe \code{_id}) : \code{F} par défaut}
}
\description{
Cette fonction permet de reformater les id des maîtres d'ouvrage en  libellé
}
\examples{
formatage.mo.libelle(data, colonne_entree = "tpswprj_client")
formatage.mo.libelle(data, colonne_entree = "prjlst_structure", conservation = T)
}
