% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.noms_propres.R
\name{formatage.noms_propres}
\alias{formatage.noms_propres}
\title{Reformatage de noms propres mal orthographiés}
\usage{
formatage.noms_propres(
  data,
  Operation = c("Nettoyage", "Développement"),
  ColonneEntree = NA_character_,
  ColonneSortie = NA_character_
)
}
\arguments{
\item{data}{Dataframe contenant les données à transformer}

\item{Operation}{Type de transformation que l'on souhaite réaliser : \code{Nettoyage} par défaut (sortie sans article), \code{Développement} possible (sortie avec article)}

\item{ColonneEntree}{Champ contenant la donnée d'entrée}

\item{ColonneSortie}{Champ recevant la donnée de sortie (peut être identique au champs d'entrée si on le souhaite)}
}
\description{
Cette fonction permet de reformater des noms propres (écosystèmes, communes, etc.) mal orthographiés
}
\examples{
formatage.noms_propres(data, Operation = "Nettoyage", ColonneEntree = "chsta_diccor_valeur_incorrecte", ColonneSortie = "chsta_ecosysteme")
formatage.noms_propres(data, Operation = "Développement", ColonneEntree = "chsta_diccor_valeur_incorrecte", ColonneSortie = "chsta_ecosysteme")
}
