% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.personnel.id.R
\name{formatage.personnel.id}
\alias{formatage.personnel.id}
\title{Reformatage des prénoms/noms en id}
\usage{
formatage.personnel.id(data, colonne_entree = NA_character_, conservation = F)
}
\arguments{
\item{data}{Dataframe contenant les données à transformer}

\item{colonne_entree}{Champ contenant la donnée à ré-écrire}

\item{conservation}{Conservation du champs colonne_entree dans la sortie (avec suffixe \code{_libelle}) : \code{F} par défaut}
}
\description{
Cette fonction permet de reformater les prénoms/noms du personnel en id
}
\examples{
formatage.personnel.id(data, colonne_entree = "tpswot_personnel")
formatage.personnel.id(data, colonne_entree = "tpswot_personnel", conservation = T)
}
