% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.personnel.prenom.R
\name{formatage.personnel.prenom}
\alias{formatage.personnel.prenom}
\title{Reformatage des id en prénom}
\usage{
formatage.personnel.prenom(data, colonne_entree = NA_character_)
}
\arguments{
\item{data}{Dataframe contenant les données à transformer}

\item{colonne_entree}{Champ contenant la donnée à ré-écrire}
}
\description{
Cette fonction permet de reformater les id du personnel en prénom
}
\examples{
formatage.personnel.prenom(data, colonne_entree = "tpswot_personnel")
formatage.personnel.prenom(data, colonne_entree = "tpswot_personnel", conservation = T)
}
