% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.sapl.id.R
\name{formatage.sapl.id}
\alias{formatage.sapl.id}
\title{Reformatage des SAPL en id}
\usage{
formatage.sapl.id(data, colonne_entree = NA_character_, conservation = F)
}
\arguments{
\item{data}{Dataframe contenant les données à transformer}

\item{colonne_entree}{Champ contenant la donnée à ré-écrire}

\item{conservation}{Conservation du champs colonne_entree dans la sortie (avec suffixe \code{_libelle}) : \code{F} par défaut}
}
\description{
Cette fonction permet de reformater les libellé des SAPL en id
}
\examples{
formatage.sapl.id(data, colonne_entree = "sapl")
formatage.sapl.id(data, colonne_entree = "sapl", conservation = T)
}
