% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.sapl.R
\name{formatage.sapl}
\alias{formatage.sapl}
\title{Ré-encodage des SAPL}
\usage{
formatage.sapl(data = data)
}
\arguments{
\item{data}{dataframe contenant au moins une colonne dont l'intitulé contient sapl}
}
\description{
Cette fonction permet de ré-encoder les noms des SAPL
}
\examples{
formatage.sapl(data)
}
\keyword{formatage}
