% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.emergence.R
\name{general.emergence}
\alias{general.emergence}
\title{Extraction observations émergence de projets}
\usage{
general.emergence(export = FALSE, exportventile = FALSE, piecejointes = FALSE)
}
\arguments{
\item{export}{Un export sous forme de table SIG est réalisé si \code{TRUE}. \code{FALSE} par défault.}

\item{exportventile}{Export complémentaire sous forme de ventilation thématique si \code{TRUE}. \code{FALSE} par défault.}

\item{piecejointes}{Extraction des pièces jointes si \code{TRUE}. \code{FALSE} par défault.}
}
\description{
Cette fonction permet d'extraire les données des observations réalisées en vue de l'émergence de projets
}
\examples{
general.emergence(export = TRUE)
general.emergence(piecejointes = TRUE)
general.emergence(export = TRUE, piecejointes = TRUE)
general.emergence(export = TRUE, exportventile = TRUE, piecejointes = TRUE)
}
