% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.identification.R
\name{general.identification}
\alias{general.identification}
\title{Identification de la machine}
\usage{
general.identification()
}
\description{
Cette fonction permet d'identifier la machine sur laquelle le code est exécuté
}
