% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.mail.R
\name{general.mail}
\alias{general.mail}
\title{Envoi de mail}
\usage{
general.mail(
  destinataire = NA_character_,
  objet = NA_character_,
  message = NA_character_,
  expediteur = NA_character_,
  expediteur_mdp = NA_character_
)
}
\arguments{
\item{destinataire}{Destinataire du mail}

\item{objet}{Objet à inclure}

\item{message}{Message à inclure}

\item{expediteur}{Expéditeur du mail}

\item{expediteur_mdp}{Mot de passe du compte d'expédition}
}
\description{
Cette fonction permet d'envoyer des mails
}
