% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrographie.SDAGE.R
\name{hydrographie.SDAGE}
\alias{hydrographie.SDAGE}
\title{Extraction du réseau hydrographique selon le SDAGE}
\usage{
hydrographie.SDAGE(export = FALSE)
}
\arguments{
\item{export}{Forme du dataframe de export - \code{FALSE} (par défault) ou \code{TRUE}}
}
\description{
Cette fonction permet d'extraire les masses d'eau au sens du SDAGE
}
\examples{
hydrographie.SDAGE()
MDO <- hydrographie.SDAGE()
Resultats <- hydrographie.SDAGE(export = "TRUE")
}
\keyword{donnees}
