% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrovu.collecte.R
\name{hydrovu.collecte}
\alias{hydrovu.collecte}
\title{Collecte des données de l'API HydroVu}
\usage{
hydrovu.collecte(
  id = NA_character_,
  chsta_coderhj = NA_character_,
  modem = NA_character_,
  capteur = NA_character_,
  time_end = now(),
  profondeur = 10,
  intervalle = 6
)
}
\arguments{
\item{id}{Id du capteur recherché sur l'API}

\item{chsta_coderhj}{Station concernées par l'\code{id} recherché sur l'API}

\item{modem}{Modem concerné par l'\code{id} recherché sur l'API}

\item{capteur}{Capteur concerné par l'\code{id} recherché sur l'API}

\item{time_end}{Timecode de fin de recherche (\code{now()} par défaut)}

\item{profondeur}{Profondeur temporelle recherchée (en jours)}

\item{intervalle}{Intervalle entre deux séquences de collecte (en heures)}
}
\description{
Cette fonction permet de collecter les données disponibles l'API HydroVu
}
\examples{
hydrovu.collecte("5327553842264847", "BONlac", "852311", "794849")
}
\keyword{chronique}
