% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam.resultats.R
\name{iam.resultats}
\alias{iam.resultats}
\title{Extraction des résultats d'IAM}
\usage{
iam.resultats(
  ListeStations = data.frame(Nom = character(0)),
  Sortie = c("Simple", "Propre", "Complet"),
  periode = c("10ans", "5ans", "20ans", "Complet", "4campagnes")
)
}
\arguments{
\item{ListeStations}{Dataframe contenant un colonne "Nom" avec le code de la station (RHJ)}

\item{Sortie}{Forme du dataframe de sortie - \code{Simple} (par défault), \code{Propre} (format diffusable, avec stations) ou \code{Complet} (tous les champs)}

\item{periode}{Permet de limiter la durée des données traitées : 5ans, 10ans (par défaut), 20ans, Complet, 4campagnes}
}
\description{
Cette fonction permet de collecter les résultats des IAM
}
\examples{
iam.resultats()
Resultats <- iam.resultats()
Resultats <- iam.resultats(Sortie = "Complet")
iam.resultats(data.frame(Nom = "SOR10-2"), Sortie = "Propre", periode = "20ans")
iam.resultats(Sortie = "Complet") \%>\% st_join(stations.territoire("Commune", c("Montmorot"))) \%>\% filter(!is.na(tpcomm_commune_libellesansarticle))
}
\keyword{donnees}
