% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.captures.R
\name{MI.captures}
\alias{MI.captures}
\title{Extraction des données des captures de MI}
\usage{
MI.captures(id_operation = NA_integer_, sortie = c("Brute", "Complète"))
}
\arguments{
\item{id_operation}{id de l'opération recherchée}

\item{sortie}{Forme du dataframe de sortie - \code{Brute} (par défault) ou \code{Complète} (avec données des prélèvements et de station)}
}
\description{
Récupère les données de captures d'une opération de suivi MI
}
\examples{
MI.captures()
MI.captures(12)
c(12, 14) \%>\% purrr::map_dfr(~ MI.captures(.))
captures <- operations \%>\% group_split(id) \%>\% purrr::map_dfr(~ MI.captures(.$id))
c("GLA5-5", "GLA6-2") \%>\% purrr::map_dfr(~ MI.operations(., id_operation = T)) \%>\% group_split(id) \%>\% purrr::map_dfr(~ MI.captures(.$id))
}
\keyword{donnees}
