% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.operations.R
\name{MI.operations}
\alias{MI.operations}
\title{Listing des opérations (MI) pour une station}
\usage{
MI.operations(
  code_station = NA_character_,
  id_operation = F,
  sortie = c("Simple", "Propre", "Complet")
)
}
\arguments{
\item{code_station}{Code RHJ de la station}

\item{id_operation}{Affichage de l'id des opérations - \code{FALSE} (par défault)}

\item{sortie}{Forme du dataframe de sortie - \code{Simple} (par défault), \code{Propre} (format diffusable, avec stations) ou \code{Complet} (tous les champs)}
}
\description{
Cette fonction permet de lister les différentes opérations pour une liste donnée de stations
}
\examples{
MI.operations()
MI.operations(id_operation = TRUE)
MI.operations("BCH5-7")
MI.operations("BCH5-7", id_operation = TRUE)
MI.operations("BCH5-7", sortie = "Complet")
c("GLA5-5", "GLA6-2") \%>\% purrr::map_dfr(~ MI.operations(., sortie = "Propre", id_operation = T))
captures <- Stations \%>\% group_split(chsta_coderhj) \%>\% purrr::map_dfr(~ MI.operations(.$chsta_coderhj, sortie = "Propre"))
}
\keyword{macro-invertébrés}
