% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.systematique.presence.R
\name{MI.systematique.presence}
\alias{MI.systematique.presence}
\title{Vérification existence MI}
\usage{
MI.systematique.presence(data)
}
\arguments{
\item{data}{Jeu de données à vérifier, issu de MI.captures}
}
\description{
Cette fonction permet de vérifier la présence des taxons de MI
}
\examples{
MI.systematique.presence(data)
}
\keyword{MI}
