% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentime.association.personnel.R
\name{opentime.association.personnel}
\alias{opentime.association.personnel}
\title{Association du personnel aux données OpenTime}
\usage{
opentime.association.personnel(
  data = NA_character_,
  format = c("libelle_separe", "libelle_complet", "id")
)
}
\arguments{
\item{data}{Jeu de données à traiter, issu de \code{opentime.tps.attendu()} par exemple}

\item{format}{Format de la sortie : \code{libelle_separe} (par défaut), \code{libelle_complet} ou \code{id}}
}
\description{
Associe proprement les personnels aux données issues de OpenTime
}
\examples{
"monfichier.xlsx" \%>\% opentime.tps.attendu() \%>\% opentime.association.personnel() \%>\% opentime.projet()
}
\keyword{personnel}
