% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentime.ouverture.R
\name{opentime.ouverture}
\alias{opentime.ouverture}
\title{Ouverture des fichiers excels de données détaillées}
\usage{
opentime.ouverture(fichier = NA_character_)
}
\description{
Ouvre les données excel à partir des données fines téléchargées depuis OpenTime et les reformate
}
\examples{
"monfichier.xlsx" \%>\% opentime.ouverture \%>\% BDD.format(Type = "Temps de travail") \%>\% opentime.projet()
}
\keyword{personnel}
