% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentime.projet.R
\name{opentime.projet}
\alias{opentime.projet}
\title{Extraction des identifiants de projet issus des données OpenTime téléchargées}
\usage{
opentime.projet(data = NA_character_)
}
\description{
Ajoute une colonne \code{projet_id} à partir des données fines téléchargées depuis OpenTime
}
\examples{
donnees_telechargees \%>\% BDD.format(Type = "Temps de travail") \%>\% opentime.projet()
}
\keyword{personnel}
