% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentime.tps.attendu.R
\name{opentime.tps.attendu}
\alias{opentime.tps.attendu}
\title{Ouverture données OpenTime "détails prévisionnel"}
\usage{
opentime.tps.attendu(fichier = NA_character_)
}
\arguments{
\item{fichier}{Fichiers \code{.xlsx} issu de OpenTime à traiter (\code{détails prévisonnels})}
}
\description{
Ouvre les données "détails temps passé" OpenTime et les reformate
}
\examples{
"monfichier.xlsx" \%>\% opentime.tps.attendu() \%>\% opentime.association.personnel() \%>\% BDD.format(Type = "Temps de travail") \%>\% opentime.projet()
}
\keyword{personnel}
