% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentime.tps.realise.R
\name{opentime.tps.realise}
\alias{opentime.tps.realise}
\title{Ouverture données OpenTime "détails temps passé"}
\usage{
opentime.tps.realise(fichier = NA_character_)
}
\arguments{
\item{fichier}{Fichiers \code{.xlsx} issu de OpenTime à traiter (\code{détails temps passé})}
}
\description{
Ouvre les données "détails temps passé" OpenTime et les reformate
}
\examples{
"monfichier.xlsx" \%>\% opentime.tps.realise() \%>\% BDD.format(Type = "Temps de travail") \%>\% opentime.projet()
}
\keyword{personnel}
