% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.mesures.R
\name{PC.mesures}
\alias{PC.mesures}
\title{Collecte des mesures de physico-chimie}
\usage{
PC.mesures(station = "", date = "", sortie = c("Propre", "Simple", "Complet"))
}
\arguments{
\item{station}{Station recherchée (pas de filtrage si vide)}

\item{date}{Date recherchée (pas de filtrage si vide)}

\item{sortie}{Forme du dataframe de sortie - \code{Propre} (par défault) ou \code{Simple} ou \code{Complet}}
}
\description{
Cette fonction permet de collecter les mesures de physico-chimie
}
\examples{
PC.mesures()
PC.mesures(station = "SOR10-2")
PC.mesures("SOR10-2", sortie = "Propre")
PC.mesures("Lac de Chalain", date = "2010-01-01", sortie = "Propre")
operations_a_corriger \%>\% group_split(cle) \%>\% map(~ PC.mesures(station = .$pcsvi_coderhj, date = .$pcsvi_date, sortie = "Complet")) \%> list_rbind()
}
\keyword{pc}
