% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.operations.R
\name{PC.operations}
\alias{PC.operations}
\title{Collecte des opérations de physico-chimie}
\usage{
PC.operations(
  station = "",
  date = "",
  sortie = c("Propre", "Simple", "Complet")
)
}
\arguments{
\item{station}{Station recherchée (pas de filtrage si vide)}

\item{date}{Date recherchée (pas de filtrage si vide)}

\item{sortie}{Forme du dataframe de sortie - \code{Propre} (par défault) ou \code{Simple} ou \code{Complet}}
}
\description{
Cette fonction permet de collecter les opérations de physico-chimie
}
\examples{
PC.operations()
PC.operations(station = "SOR10-2")
PC.operations("SOR10-2", sortie = "Propre")
PC.operations("Lac de Chalain", date = "2010-01-01", sortie = "Propre")
}
\keyword{pc}
