% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personnel.formatAC.R
\name{personnel.formatAC}
\alias{personnel.formatAC}
\title{Extraction et réencodage données OpenTime pour AC AE}
\usage{
personnel.formatAC(data, projet = NA_character_)
}
\arguments{
\item{data}{Jeu de données à formater}

\item{projet}{Accord-cadre concerné}
}
\description{
Permet l'extraction et le réencodage des données OpenTime pour les accords-cadre avec l'Agence de l'eau
}
\examples{
personnel.formatAC(data, "Convention-cadre AERMC-FD39-2020")
}
\keyword{personnel}
