% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personnel.projection.R
\name{personnel.projection}
\alias{personnel.projection}
\title{Adjonction des projections de réalisation annuelles}
\usage{
personnel.projection(data = NA_character_, type = NA_character_)
}
\arguments{
\item{data}{Jeu de données en entrée, issu du temps de travail théorique de \code{personnel.projet}}

\item{type}{Type de projection à conserver \code{vide} par défaut <-> pas de filtrage}
}
\description{
Cette fonction permet d'adjoindre au données théoriques de réalisation une projection temporelle annuelle
}
\examples{
personnel.projection(tps_travail_theorique)
personnel.projection(tps_travail_theorique, type = "Brouillon")
personnel.projection(tps_travail_theorique, type = c("Brouillon", "Brouillon 2"))
}
\keyword{personnel}
