% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.programmation.R
\name{poisson.programmation}
\alias{poisson.programmation}
\title{Extraction des suivis piscicoles programmés}
\usage{
poisson.programmation(
  annee = NA_integer_,
  projet_id = NA_integer_,
  geojson = T,
  excel = T,
  shp = F,
  kml = F
)
}
\arguments{
\item{annee}{Année recherchée}

\item{projet_id}{Projet recherché (facultatif)}

\item{geojson}{Export au format geojson (\code{true} par défaut)}

\item{excel}{Export au format excel (\code{true} par défaut)}

\item{shp}{Export au format shapefile (\code{false} par défaut)}

\item{kml}{Export au format kml (\code{false} par défaut)}
}
\description{
Cette fonction permet de réaliser une extraction des suivis piscicoles programmés, à partir de la table fd_production.poissons_programmation
}
\examples{
poisson.programmation(2022)
poisson.programmation(2022, 224)
poisson.programmation(2022, 224, geojson = F, excel = F)
}
