% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.PV.R
\name{poissons.PV}
\alias{poissons.PV}
\title{Exportation des PV}
\usage{
poissons.PV(ecosysteme = "")
}
\arguments{
\item{ecosysteme}{Code de l'écosystème. Si vide (par défaut), exporte l'intégralité du jeu de données}
}
\description{
Cette fonction permet d'exporter les données de PV de gestion piscicole
}
\examples{
poissons.PV()
poissons.PV("ROU")
}
\keyword{poissons}
