% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.resultats.R
\name{poissons.resultats}
\alias{poissons.resultats}
\title{Extraction des résultats pour une station}
\usage{
poissons.resultats(
  ListeStations = data.frame(Nom = character(0)),
  Sortie = c("Simple", "Propre", "Complet"),
  periode = c("10ans", "20ans", "Complet", "4campagnes")
)
}
\arguments{
\item{ListeStations}{Dataframe contenant un colonne "Nom" avec le code de la station (RHJ)}

\item{Sortie}{Forme du dataframe de sortie - \code{Simple} (par défault), \code{Propre} (format diffusable, avec stations) ou \code{Complet} (tous les champs)}

\item{periode}{Permet de limiter la durée des données traitées : 10ans (par défaut), 20ans, Complet , 4campagnes}
}
\description{
Cette fonction permet de collecter les différents résultats pour une liste donnée de stations
}
\examples{
poissons.resultats()
Resultats <- poissons.resultats()
poissons.resultats(data.frame(Nom = "SOR10-2"), Sortie = "Propre")
}
\keyword{donnees}
