% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.traitement.R
\name{poissons.traitement}
\alias{poissons.traitement}
\title{Traitement semi-automatisé de données piscicoles}
\usage{
poissons.traitement(
  data = data,
  export = F,
  commentaires = F,
  archivage = c("Aucun", "Partiel", "Complet")
)
}
\arguments{
\item{data}{Jeu de données en entrée (captures issues de poissons.captures)}

\item{export}{\code{FALSE} par défault. Permet d'exporter les données}

\item{commentaires}{\code{FALSE} par défault. Permet d'extraire le commentaire associé aux opérations}

\item{archivage}{Si Aucun (par défault), ne va pas créer une archives .zip du répertoire de sortie. Si Partiel, créé une archive et conserve le répertoire. Si Complet, créé une archive et supprimer le répertoire.}
}
\description{
Cette fonction permet de traiter de manière semi-automatisée des données piscicoles (histogrammes, graphiques CAA/CAR, etc.)
}
\examples{
poissons.traitement(data)
poissons.traitement(data, export=T)
poissons.captures("AIN18-4") \%>\% poissons.traitement(export=T)
}
\keyword{poissons}
