% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet.calculInitial.R
\name{projet.calculInitial}
\alias{projet.calculInitial}
\title{Calcul initial des coûts de projet}
\usage{
projet.calculInitial(nom_projet)
}
\arguments{
\item{nom_projet}{Nom du dataframe contenant les données "attendues" à traiter, dans le format de tpstravail_recapitulatif}
}
\description{
Cette fonction permet de calculer les coûts d'un projet au moment de son montage. Elle s'appuie sur le format saisi dans la table tpstravail_recapitulatif et réalise les conversions à partir des volumes unitaires, et ordonne les sujets
}
\examples{
recap_data_to_add <- projet.calculInitial(nom_projet)
tbl(dbD, dbplyr::in_schema("fd_production", "tpstravail_recapitulatif")) \%>\% filter(tpswrecap_projet == 212) \%>\% collect() \%>\% projet.calculInitial()
}
