% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet.comptabilite.R
\name{projet.comptabilite}
\alias{projet.comptabilite}
\title{Affectation des projet à la comptabilité}
\usage{
projet.comptabilite(fichier = NA_character_)
}
\arguments{
\item{fichier}{Fichier txt à traiter, extrait du logiciel Ciel Compta (grand livre analytique)}
}
\description{
Cette fonction permet d'affecter aux opérations comptables les projets associés par nettoyage
}
\examples{
"2024 Grand livre analytique.txt" \%>\% projet.comptabilite()
}
