% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet.extraction.R
\name{projet.extraction}
\alias{projet.extraction}
\title{Extraction des coûts de projet}
\usage{
projet.extraction(projet_id)
}
\arguments{
\item{projet_id}{Id du projet concerné, à partir de la table fd_production.tpstravail_projets}
}
\description{
Cette fonction permet d'extraire les coûts d'un projet sous forme excel et png au moment de son montage. Elle s'appuie sur le format saisi dans la table tpstravail_recapitulatif et réalise les conversions à partir des volumes unitaires, et ordonne les sujets
}
\examples{
projet.extraction(96)
Projets \%>\% group_split(id) \%>\% purrr::map(~ projet.extraction(.$id))
}
