% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet.suivi.R
\name{projet.suivi}
\alias{projet.suivi}
\title{Tableau de suivi des projets}
\usage{
projet.suivi(
  opentime_attendu = NA_character_,
  opentime_realise = NA_character_,
  filtrage_periode_projets_clos = 24,
  filtrage_discussion = FALSE,
  save = T
)
}
\arguments{
\item{opentime_attendu}{Fichier OpenTime des détails prévisionnel}

\item{opentime_realise}{Fichier OpenTime des détails temps passé}

\item{filtrage_periode_projets_clos}{Durée (en mois) de conservation des projets soldés}

\item{filtrage_discussion}{Si \code{FALSE} (par défaut), on conserve les projets "En développement|En discussion|En engagement"}

\item{save}{Si \code{TRUE} (\code{FALSE} par défaut), exportation et ouverture du suivi en fichier xlsx}
}
\description{
Cette fonction permet d'établir le tableau de suivi des projets
}
\examples{
projet.suivi("/Users/jean-baptistefagot/Downloads/export_détails_prévisionnel_20260311.xlsx", "/Users/jean-baptistefagot/Downloads/export_détails_temps_passé_20260311.xlsx")
previsionnel <- "/Users/jean-baptistefagot/Nextcloud/NAS-FD/FD39/Activité/Suivi_activité/opentime_prévisionnels/" \%>\% dir_ls() \%>\% last() ; realise <- "/Users/jean-baptistefagot/Nextcloud/NAS-FD/FD39/Activité/Suivi_activité/opentime_temps_réels/" \%>\% dir_ls() \%>\% last(); projet.suivi(previsionnel, realise)
}
