% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.recuperation.R
\name{table.recuperation}
\alias{table.recuperation}
\title{Collecte/export d'une table en base de données}
\usage{
table.recuperation(
  table = NA_character_,
  condition = NA_character_,
  export = FALSE
)
}
\arguments{
\item{table}{Nom de la table à exporter, sans le schéma}

\item{condition}{Éventuelle condition de filtrage de la table à collecter/exporter}

\item{export}{Exportation de la table (\code{FALSE} par défaut)}
}
\description{
Cette fonction permet de collecter/exporter une table en base de données
}
\examples{
table.recuperation("gis_emprise", export = T)
table.recuperation("chroniques_resultats")
table.recuperation("chroniques_resultats", export = T)
table.recuperation("topographie_operations_habitats", "id = 12")
c("gis_emprise", "pecheloisir_categories_polygones", "hydrographie_contextespdpg") \%>\% map(~ table.recuperation(., export = T))
}
\keyword{cartographie}
