% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermie.spatiale.mesures.R
\name{thermie.spatiale.mesures}
\alias{thermie.spatiale.mesures}
\title{Extraction de données de thermie spatialisées}
\usage{
thermie.spatiale.mesures(thspaop_id = NA_character_)
}
\arguments{
\item{thspaop_id}{Identifiant du projet}
}
\description{
Cette fonction permet d'extraire les données de thermies spatiales (IRTa par exemple) sous forme de ponctuels
}
\examples{
mesures <- thermie.spatiale.mesures()
mesures <- thermie.spatiale.mesures("1")
mesures <- c(1, 2) \%>\% tibble(id = .) \%>\% group_split(id) \%>\% map(~ thermie.spatiale.mesures(.)) \%>\% reduce(rbind)
mesures <- thermie.spatiale.mesures("1") \%>\% filter(thspapo_type_valeur == "Lissage par ondelette sur le linéaire") \%>\% mapview::mapview(zcol = "thspapo_t_med")
mesures <- thermie.spatiale.mesures("2") \%>\% filter(thspapo_type_valeur == "Centroïde de patch froid") \%>\% mapview::mapview(zcol = "thspapo_a_med", cex = "thspapo_superficie")
}
\keyword{thermie}
