% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.habitats.R
\name{topographie.habitats}
\alias{topographie.habitats}
\title{Extraction de données d'habitats}
\usage{
topographie.habitats(tphabop_id = NA_character_)
}
\arguments{
\item{tphabop_id}{Identifiant de l'opération}
}
\description{
Cette fonction permet d'extraire les données de suivis d'habitats
}
\examples{
mesures <- topographie.habitats()
mesures <- topographie.habitats("12")
mesures <- c(17, 23) \%>\% tibble(id = .) \%>\% group_split(id) \%>\% map(~ topographie.habitats(.)) \%>\% reduce(rbind)
}
\keyword{topographie}
