% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.operations.habitats.R
\name{topographie.operations.habitats}
\alias{topographie.operations.habitats}
\title{Extraction de données d'opérations de suivi des habitats}
\usage{
topographie.operations.habitats(id = NA_character_, tphabop_id = NA_character_)
}
\arguments{
\item{id}{Éventuel identifiant de l'opération}

\item{tphabop_projet_id}{Éventuel identifiant du projet}
}
\description{
Cette fonction permet d'extraire les données des opérations de suivis d'habitats
}
\examples{
mesures <- topographie.operations.habitats()
mesures <- topographie.operations.habitats("12")
mesures <- c(17, 23) \%>\% tibble(id = .) \%>\% group_split(id) \%>\% map(~ topographie.operations.habitats(.)) \%>\% reduce(rbind)
}
\keyword{topographie}
