% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.traitement.R
\name{chronique.traitement}
\alias{chronique.traitement}
\title{Exécuter le traitement de données de chroniques}
\usage{
chronique.traitement(
  data = data,
  projet = NA_character_,
  export = T,
  exportfigures = T,
  style = c("boxplot", "violon"),
  exportfigures_chronique_classique = T,
  exportfigures_cumul_degresjours = T,
  exportfigures_boxplot_interannuel = T,
  exportfigures_classes_calendaires = T,
  exportfigures_profil_longitudinal = T,
  exportfigures_preferendums_especes = T,
  exportDCE = F,
  filtrage = F,
  filtrageanneevmm = TRUE,
  filtragedatefperiode = "07-15",
  filtragenbj = 75,
  seuils = c(25, 22, 19, 15, 4),
  seuilexcesdefaut = 12,
  dep39 = c(FALSE, TRUE, "autre"),
  localisation_stations = NA,
  localisation_commentaires = NA,
  localisation_suiviterrain = NA,
  localisation_capteurs = NA,
  localisation_donnees = NA,
  archivage = c("Aucun", "Partiel", "Complet"),
  log = c("Simple", "Verbeux", "Aucun")
)
}
\arguments{
\item{data}{Data.frame issu de chronique.mesures, pouvant contenir différentes stations}

\item{projet}{Nom du projet}

\item{export}{Si \code{TRUE} (par défaut), exporte les résultats/figures. Si \code{FALSE}, ne les exporte pas.}

\item{exportfigures}{Si \code{TRUE} (par défaut), exporte les figures. Si \code{FALSE}, ne les exporte pas.}

\item{style}{\code{boxplot} ou \code{violon}}

\item{exportfigures_chronique_classique}{Si \code{TRUE} (par défaut), exporte les vues annuelles des chroniques. Si \code{FALSE}, ne les exporte pas.}

\item{exportfigures_cumul_degresjours}{Si \code{TRUE} (par défaut), exporte les cumuls de degrés-jours. Si \code{FALSE}, ne les exporte pas.}

\item{exportfigures_boxplot_interannuel}{Si \code{TRUE} (par défaut), exporte les boxplots interannuels. Si \code{FALSE}, ne les exporte pas.}

\item{exportfigures_classes_calendaires}{Si \code{TRUE} (par défaut), exporte les vues sous forme de classes calendaires. Si \code{FALSE}, ne les exporte pas.}

\item{exportfigures_profil_longitudinal}{Si \code{TRUE} (par défaut), exporte les figures de profils longitudinaux. Si \code{FALSE}, ne les exporte pas.}

\item{exportfigures_preferendums_especes}{Si \code{TRUE} (par défaut), exporte les figures de préférendums des espèces avec les Vmm30j. Si \code{FALSE}, ne les exporte pas.}

\item{exportDCE}{Si \code{TRUE}, exporte les mesures au format DCE. Si \code{FALSE} (par défaut), ne les exporte pas.}

\item{filtrage}{Si \code{TRUE}, filtre les résultats avec chronique.resultats.filtrage pour ne conserver que les pertinents pour les représentations graphiques inter-annuelles/inter-sites. \code{FALSE} (par défaut)}

\item{filtrageanneevmm}{Si \code{TRUE} (par défaut), ne conserve que les résultats dont l'année de fin de vmm30j est égale à l'année biologique du résultat}

\item{filtragedatefperiode}{Ne conserver que les résultats dont la date de fin de période est postérieure à la date choisie ("07-15", soit le 15 juillet par défaut)}

\item{filtragenbj}{Nombre de journées minimales pour chaque résultat à conserver (75 par défaut).}

\item{seuils}{Seuils de valeurs (25,22,19,15,4 par défaut) dont il faut tester les dépassements. L'ordre de sortie correspond à l'ordre de cette liste.}

\item{seuilexcesdefaut}{Seuil en-deça duquel les dépassements sont testés par défaut, sinon ils sont testés par excès (valeur de 12 par défaut).}

\item{dep39}{Si \code{FALSE} (par défaut), ne va pas rechercher les données de stations dans la base locale et donc export simplifié. Si \code{TRUE}, fait la jointure SIG. Possibilité d'utiliser \code{autre} afin de sélectionner un fichier source de stations}

\item{localisation_stations}{Si \code{NA} (par défaut), ouvre un applet pour localiser le fichier des stations}

\item{localisation_commentaires}{Si \code{NA} (par défaut), ouvre un applet pour localiser le fichier des commentaires}

\item{localisation_suiviterrain}{Si \code{NA} (par défaut), ouvre un applet pour localiser le fichier de suivi de terrain}

\item{localisation_capteurs}{Si \code{NA} (par défaut), ouvre un applet pour localiser le fichier des capteurs}

\item{localisation_donnees}{Si \code{NA} (par défaut), ouvre un applet pour localiser le fichier des données brutes}

\item{archivage}{Si \code{Aucun} (par défaut), ne va pas créer une archives .zip du répertoire de sortie. Si \code{Partiel}, créé une archive et conserve le répertoire. Si \code{Complet}, créé une archive et supprimer le répertoire.}

\item{log}{Si \code{Simple} (par défaut), le log ne sera pas très bavard. Si \code{Verbeux}, le log incluera tous les retours du tidyverse. Si \code{Aucun}, absence de log}
}
\description{
Permet d'exécuter plusieurs opérations de traitement des données de chronique d'une station
}
\examples{
chronique.traitement(data)
chronique.traitement(data, export = F)
DataTravail <- chronique.mesures("SUR0-9", "Thermie") \%>\% chronique.traitement()
DataTravail <- chronique.mesures("SUR0-9", "Thermie") \%>\% chronique.traitement(., export = F, filtrage = T)
DataTravail <- tbl(db,"Mesures") \%>\% filter(CodeRDT == "DOU393-2" | CodeRDT == "DOU394-5") \%>\% collect() \%>\% chronique.traitement()
}
\keyword{chronique}
