% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.personnel.prenom.R
\name{formatage.personnel.prenom}
\alias{formatage.personnel.prenom}
\title{Reformatage des id en prénom}
\usage{
formatage.personnel.prenom(
  data,
  colonne_entree = NA_character_,
  conservation = F
)
}
\arguments{
\item{data}{Dataframe contenant les données à transformer}

\item{colonne_entree}{Champ contenant la donnée à ré-écrire}

\item{conservation}{Conservation du champs colonne_entree dans la sortie (avec suffixe \code{_id}) : \code{F} par défaut}
}
\description{
Cette fonction permet de reformater les id du personnel en prénom
}
\examples{
formatage.personnel.prenom(data, colonne_entree = "tpswot_personnel")
formatage.personnel.prenom(data, colonne_entree = "tpswot_personnel", conservation = T)
}
