% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam.resultats.R
\name{iam.resultats}
\alias{iam.resultats}
\title{Extraction des résultats d'IAM}
\usage{
iam.resultats(tpiam_op_id = NA_character_)
}
\arguments{
\item{tpiam_op_id}{Identifiant de l'opération}
}
\description{
Cette fonction permet de collecter les résultats des IAM
}
\examples{
mesures <- iam.resultats()
mesures <- iam.resultats("12")
mesures <- c(17, 23) \%>\% tibble(id = .) \%>\% group_split(id) \%>\% map(~ iam.resultats(.)) \%>\% reduce(rbind)
}
\keyword{topographie}
